+++ modules/seqexec/model/src/main/scala/seqexec/model/Model.scala
@@ -10,7 +10,7 @@ import cats.implicits._
 import gem.Observation
 import seqexec.model.enum._
 
-@SuppressWarnings(Array("org.wartremover.warts.PublicInference", "org.wartremover.warts.IsInstanceOf"))
+// @SuppressWarnings(Array("org.wartremover.warts.PublicInference", "org.wartremover.warts.IsInstanceOf"))
 // scalastyle:off
 object Model {
 
@@ -39,6 +39,7 @@ object Model {
     name: String
   )
 
+@SuppressWarnings(Array("org.wartremover.warts.PublicInference", "org.wartremover.warts.IsInstanceOf"))
   object SequenceMetadata {
     implicit val eq: Eq[SequenceMetadata] =
       Eq.by(x => (x.instrument, x.observer, x.name))
@@ -52,6 +53,7 @@ object Model {
     willStopIn: Option[Int]
   )
 
+@SuppressWarnings(Array("org.wartremover.warts.PublicInference", "org.wartremover.warts.IsInstanceOf"))
   object SequenceView {
     implicit val eq: Eq[SequenceView] =
       Eq.by { x =>
@@ -59,60 +61,6 @@ object Model {
       }
   }
 
-  sealed trait OffsetAxis {
-    val configItem: String
-  }
-  object OffsetAxis {
-    case object AxisP extends OffsetAxis {
-      val configItem = "p"
-    }
-    case object AxisQ extends OffsetAxis {
-      val configItem = "q"
-    }
-    implicit val show: Show[OffsetAxis] = Show.show {
-      case AxisP => "p"
-      case AxisQ => "q"
-    }
-  }
-
-  sealed trait Offset {
-    val value: Double
-  }
-  object Offset {
-    implicit val equal: Eq[Offset] =
-      Eq.by {
-        case p: TelescopeOffset.P => Left(p)
-        case q: TelescopeOffset.Q => Right(q)
-      }
-
-    def Zero(axis: OffsetAxis): Offset = axis match {
-      case OffsetAxis.AxisP => TelescopeOffset.P.Zero
-      case OffsetAxis.AxisQ => TelescopeOffset.Q.Zero
-    }
-  }
-
-  // Telescope offsets, roughly based on gem
-  final case class TelescopeOffset(p: TelescopeOffset.P, q: TelescopeOffset.Q)
-  object TelescopeOffset {
-    /** P component of an angular offset.. */
-    final case class P(value: Double) extends Offset
-    object P {
-      val Zero: P = P(0.0)
-      implicit val order: Order[P] = Order.by(_.value)
-
-    }
-    /** Q component of an angular offset.. */
-    final case class Q(value: Double) extends Offset
-    object Q {
-      val Zero: Q = Q(0.0)
-      implicit val order: Order[Q] = Order.by(_.value)
-
-    }
-    implicit val eq: Eq[TelescopeOffset] =
-      Eq.by(x => (x.p, x.q))
-
-    implicit val show: Show[TelescopeOffset] = Show.fromToString
-  }
 
 }
 //scalastyle:on
